refactor(library/standard_kernel): move standard kernel into kernel
