+++ src/kernel/CMakeLists.txt
@@ -3,4 +3,4 @@ replace_fn.cpp free_vars.cpp abstract.cpp instantiate.cpp
 formatter.cpp declaration.cpp environment.cpp pos_info_provider.cpp
 type_checker.cpp error_msgs.cpp kernel_exception.cpp
 normalizer_extension.cpp init_module.cpp expr_cache.cpp
-equiv_manager.cpp abstract_type_context.cpp)
+equiv_manager.cpp abstract_type_context.cpp standard_kernel.cpp)
