+++ src/kernel/standard_kernel.cpp
