+++ src/shell/lean.cpp
@@ -31,7 +31,7 @@ Author: Leonardo de Moura
 #include "library/st_task_queue.h"
 #include "library/mt_task_queue.h"
 #include "library/module_mgr.h"
-#include "library/standard_kernel.h"
+#include "kernel/standard_kernel.h"
 #include "library/module.h"
 #include "library/type_context.h"
 #include "library/io_state_stream.h"
